+++ robot_tests/supporting.py
@@ -1,4 +1,10 @@
 import json
+import subprocess
+
+def run_and_return_code_output_error(command):
+    task = subprocess.Popen(command, shell=True, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    out, err = task.communicate()
+    return task.returncode, out.decode(), err.decode()
 
 def parse_json(content):
     parsed = json.loads(content)